Updated Order Pipeline
Our initial understanding of the order lifecycle was incorrect. Orders are created after payment. Before that baskets should be used to contain the items the user wants to purchase. This commit simplifies the pipeline to reflect this understanding.

Note that a much of the payment processor code is removed by this commit. An updated version of the payment processor (not currently in use) will be added in a future commit.