Add refund detail view
Adds a refund detail view to Oscar which displays information about a single refund and allows a support team member to perform operations on that refund. Also provides UI elements on the refund list view which can be used to approve or deny a refund. Credit and revocation are coupled as far as the UI is concerned.