Account for coupon_vouchers in ProductAttributeValueSerializer
The way the ProductSerializer worked is it would retrieve product attributes assuming they are not of type
entity and therefor it would return an `object is not serializable` exception in case of a coupon_voucher
product attribute. This commit includes the possible coupon_voucher attribute to the serializer fixing
this issue.