Change the default values of max_uses field in coupon form.
When the user selects 'Can be used multiple times by multiple users'
usage limitation option, the default value would be 1 which doesn't
make sense for multi-use vouchers. Instead now it's default is blank
which, if left that way, will default to 10000 which is the arbitrary
limit that Oscar sets for this attribute, and the minimum limit is set
to 2. Values of other usage limitations are left the same.

https://openedx.atlassian.net/browse/SOL-2109