Updated Travis build
The run_tests.sh script is replaced with individual steps in the Travis configuration file. This allows us to better measure the execution time of individual steps of the build, and actually fail the build when individual steps fail.

Additionally, the apt packages that were once installed by run_tests.sh are now installed in the Docker image by default.