Remove the check of the LMS heartbeat
We know that ecommerce won't fall out of the loadbalancer, but it still
puts load on the LMS (800 of the roughly 4,000 checks we see).  We're
also ratcheting down the frequency of heartbeats based on the discussion
here
https://openedx.atlassian.net/wiki/display/EdxOps/Design+for+scaling+White+Labels+and+not+drowning+in+health+checks

We forsee in the future a "is gunicorn alive" and "detailed heartbeat"
as discussed there.  At that time it might make sense to revert this and
then refactor it, but this is a much smaller/simpler change.