Change SDN check address parameter to city only.
The Treasury Dept. SDN check endpoint accepts as the address
parameter only the address or the city, but not both. Since
their documentation mentions only an example where the address
parameter is a city name, our SDN check call is changed to
send only the city as the address parameter accordingly.