Updated npm dependencies
In addition to updating the dependencies, I have removed the usage of semver modifiers. This is the second time (that I recall) these floating dependencies have been the source of broken tests on Travis, but not locally. In this case karma 1.6.0 is broken due to the issue described at https://github.com/karma-runner/karma/issues/2675.