Updated CouponView to store enterprise customer UUID in cookie
Users arriving at this view, who are associated with an enterprise customer, will have this cookie set on the base domain (e.g. .edx.org instead of ecommerce.edx.org). This cookie will be used by the LMS, in the event of a redirect for login, to properly style the page based on the associated enterprise.

ENT-367