Added switch for repeat purchase check
This check to avoid allowing learners purchasing the same item multiple times is a nuisance in development and test environments. If refund deadlines pass, there is no simple way for developers to get into a state where they can view the checkout page. Disabling this check allows for the addition of items regardless of whether they have been previously purchased.