Stop thawing baskets
Thawing the user's basket when there's a problem handling their CyberSource payment causes that basket to be re-used when making subsequent payments. This thaw can be triggered when CyberSource, seemingly unprompted, sends the service an error response soon after a successful payment.

The response contains error code 104, indicating that payment processing failed because the "merchant reference code" (i.e., order number) matches a reference code sent in the last 15 minutes. After this happens, the user's basket is left open. The next time they try to check out, the basket middleware retrieves this open basket. Although we do re-freeze the basket at this point, we will silently fail to place a new order because an order with the number corresponding to the now-frozen basket already exists. If the user is attempting to pay for something at this point, we'll charge them, but they won't receive the product they paid for and we'll show them a receipt page corresponding to their most recently placed order. Since we're re-freezing the basket, the user will be able to place a new order the next time they try to do so (the middleware will create a new basket for them).

This change prevents us from thawing the user's basket when errors occur while handling CyberSource payments. As a result, a new basket should be used for every attempted order, preventing order number collisions. This change also prevents order number collisions from causing order placement to fail silently.

LEARNER-867