Handle declined attempts to pay for existing orders
CyberSource may return an error response with reason code 104, indicating that a duplicate reference number (i.e., order number) was received in a 15 minute period. This is their way of telling us that they've declined an attempt to pay for an existing order. If this happens, we log appropriate INFO level messages and redirect the browser to the receipt page for the existing order. If CyberSource actually does agree to process a payment for an existing order, the order number collision will cause an exception to be raised on our end.

LEARNER-867