Handle voucher offer benefits without range.
For some reason we persist the association between a coupon code and an Enterprise Customer
in the Range model. Not all vouchers will have a Range associated with them. This is a quick
fix to unblock program code usage. The long term fix is to refactor the way we persist the
association between a coupon and an Enterprise Customer. See ENT-354.

ENT-507