Refactored e2e tests
The previous tests were broken and somewhat confusing to decipher. The tests are now run by pytest, rather than nose (which is no longer maintained). We are now interfacing directly with Selenium rather than going through bok-choy.

The broken tests have been removed. What remains are tests of the basic SSO integration and checkout. Usage of the course detail pages on the marketing site has been intentionally removed to (a) avoid false-negatives that result from issues with the marketing site and (b) keep the tests focused solely on the E-Commerce Service.

LEARNER-509