Updated Program Conditional Offer logic to allow option for partial programs
ProgramCourseRunSeatsCondition model's is_satisfied method checks enrollments and items in basket for condition satisfaction, allowing for discount application on remaining courses in a program when a user is already enrolled in some of the program courses. Checking for enrollments can be turned on or off in site configuration, and is defaulted to not include enrollments in conditional offer determination.

[LEARNER-1901 and LEARNER-2040]