Deprecated optional.txt
The New Relic package has been upgraded and moved to production.txt. As we move toward consolidating Ansible configuration, we are adopting best practices from more recent IDAs here in this repo. We no longer use optional.txt. The cost of installing the optional package is minimal relative to the administrative overhead of optionally installing it, thus it will always be installed. However, the code will not be active on servers that are not running the New Relic agent.

LEARNER-818