Enterprise Offers
This PR refactors the way we provide automated discounts for learners
linked to an EnterpriseCustomer. We now use a custom Range and custom
Condition for determining whether or not an offer should be applied to
a given basket. The custom range determines if the contents of the basket
are contained within the specified EnterpriseCustomer's EntepriseCustomerCatalog.
The custom Condition checks to see if the learner is associated with the
specified EnterpriseCustomer and if they have provided consent to share data
from each of the products in the basket.

ENT-638