Add Enterprise offers backend.
This commit adds the business logic implementation of Enterprise
offers. An Enterprise offer consists of a custom Condition model
which checks that the user is linked to the specified EnterpriseCustomer,
the course run(s) exist in that EnterpriseCustomer's catalogs, and the
user has granted consent to share data for the course run(s).

ENT-638