Do not check for data sharing consent when checking for valid Enterprise Offers.
We originally implemented a check to make sure that the learner had
granted consent for the course runs that were in the basket when
deciding if a given Enterprise Offer should be applied to the basket.
This prevents us from accurately displaying the discounted price on
the enterprise enrollment landing pages. This PR removes the data
sharing consent check. We will rely on the Enterprise onboarding
workflow to ensure that the learner has granted consent before they
are able to purchase the course seat at the discounted price.

ENT-638