Decrease Enterprise API cache timeout from 1 hour to 5 minutes.
Caching Enterprise data for an hour significantly increases the chance
that we will end up with stale data in the cache causing Enterprise
checkout workflows to not behave as they should. Reducing the cache
timeout will decrease the likelihood of stale data while still
maintaining the performance benefits of caching.

ENT-739