Support creating entitlements
Allow creating entitlement products via the AtomicPublication API.
Just pass them in like normal seat products, with an entitlement
product class instead of a seat class.

And remove the unused /products create endpoint (that had been
intended to support this workflow, but it's not atomic, so we'll
keep using the generic publication endpoint with new support for
entitlements and drop the unused create-one-entitlement endpoint).

This reverts commit fc824bf7a6a5cfffa7486a40804778b5e0f41792 which was
itself a reversion to back out these changes while we fixed our staging
services.

LEARNER-3891