Use suppress logging in jwt_decode_handler.
Temporarily, ecommerce is delegating to the edx-drf-extensions
jwt_decode_handler, with its own fallback to allow for multiple issuers.
Logging suppression inside the edx-drf-extensions jwt_decode_handler is
required until ecommerce is down to a single issuer and can use the
edx-drf-extensions jwt_decode_handler directly.

Note: this should be reverted as part of ARCH-276.

ARCH-283