use lms_user_id from social-auth for analytics (part 1)
This refactor will grab the (LMS) user_id from social-auth extra_data
in the database, if it is available, and provide it to the analytics
code as the user tracking id.

The different user ids are also added to NewRelic (temporarily) to
help ensure all is well.  The new custom metrics are:
- lms_user_id_tracking_context (only available for LMS to ecommerce calls)
- lms_user_id_social_auth (should be None until part 2, see below)
- ecommerce_user_id

This is part 1, because part 2 will include the auth-backends upgrade
that will make the user_id available from social-auth. This upgrade has
two prerequesites:
1. A fix to LMS that enable the scope user_id to work.
2. Configuring the oauth application for ecommerce sso to have access
to the user_id scope.

ARCH-603