use lms_user_id from social-auth for analytics (part 2)
This is part 2 of a 2 part change to get the user_id from social-auth
during the oAuth2+SSO flow to the database, and available to ecommerce
for use with analytics.

Part 2 simply has the upgrade to auth-backends 2.0.0. This version of
auth-backends will actually store the user_id in the social-auth table,
making it available to the code from (part 1 #2191) that reads this
user_id and makes it available to the analytics code.

This upgrade has two pre-requisites:
1. A fix to LMS that enables the scope user_id to work at all.
2. Configuring the oauth application for ecommerce-sso to have access
to the user_id scope.

ARCH-603