enable jwt auth for add to basket
The new payment microfrontend relies on Jwt Authentication, so the
ecommerce social-auth SSO flow is a waste of time. Therefore, when
hitting /basket/add, which happens immediately before redirecting to the
payment mfe, we want it to be able to use the Jwt Cookies if it is
going to redirect to the payment mfe.