fix flaky payment page e2e test
The e2e test test_verified_seat_payment_with_credit_card_payment_page
was flaky because it tried to select the state in certain cases before
it was selectable, after switching to a country that made it selectable.

Note: the waiting code did not fix the flakiness, so resorting to
removing the ADDRESS_US test for the payment page.

ARCH-1146