add query param to coupon redeem redirect
The CouponRedeemView redirects to the Payment MFE at times when the
Payment MFE will in turn be told to redirect elsewhere (e.g. the free
basket page).  To avoid flashing the payment form, the Payment MFE will
be sent a special query param that tells it not to display the payment
form until it knows whether or not it is needed (i.e. until it gets
a response from the server).

ARCH-1101