Change tracking middleware to use two phases
This shifts the middleware's attempt to get the ga_client_id to the
process_response phase, while leaving the add_lms_user_id call in the
process_request phase.
Additionally adds temporary logging to try and identify whether this
fixes the issue and whether or not other eventing may be broken by this
change.

REV-977