ENT-2263 Added Search By Code Functionality (#2717)
This pull request is an enhancement of search functionality in Code Management Portal. As requested in the ticket (ENT-2263), enterprise admins can now make a search by code as well. The search results would help them figure out which voucher code belongs to which coupon batch and will also give them useful information about the assignment of the searched voucher.