DISCO-1604: Fixing credit seat editing (#2844)
The filter conditions that originally existed would cause
editing the credit provider field on a credit seat to create
a new credit seat. Now we just look for the seat's id and
edit that row. If not found, it will create a new seat.

Note: This logic applies for all seats in the course.