DISCO-1607: Fixing credit seat editing (#2849)
The filter conditions that originally existed would cause
editing the credit provider field on a credit seat to create
a new credit seat. Now we use the stock record sku associated
with that seat in order to edit that row. If not found,
it will create a new seat.

Note: This logic applies for all seats in the course.