preserve 30 max_length for last_name
Implements the recommended solution by Django to preserve the 30
character limit for last_name. This is necessary in preparation for the
Django 2 upgrade, because the upcoming migration to increase the length
of last_name is being faked for edx.org in Production to avoid this
large migration.

See https://docs.djangoproject.com/en/3.0/releases/2.0/#abstractuser-last-name-max-length-increased-to-150

BOM-1487