Fix transition to enterprise-catalog call (#2906)
Correct the resource name for enterprise catalogs in the new service.
This was a change that was missed as most of the old enterprise api
urls use the standard hyphenated patterns EXCEPT for
enterprise_catalogs. Additionally, apparently the waffle flag was not
yet active when tested but it is active now.