feat!: remove get_address_from_token()
The Stripe client-side payment processor was written when the
client-side had the address collected by Stripe.

Thus, the billing address was not available to the backend and needed to
query the Stripe API to get the user's billing information.

Now that we are trying to add Stripe to the Payment MFE, we will have
the user's billing information. Thus, this commit removes
get_address_from_token() as it is no longer needed.