feat: /management fulfill actions for Stripe
FulfillFrozenBaskets is a helper class for a management page in
Ecommerce (hosted at URL /management) that fulfills baskets that are
paid but stuck in queue for fulfillment, to be used if, for example,
ecommerce-worker is down.

Add Stripe support for FulfillFrozenBaskets.

Change FulfillFrozenBaskets so it does not continue fulfillment if the
billing details are not retrievable.

For Stripe, this handles when:
* The PaymentMethod was not expanded in the PaymentIntent stored in the
  PaymentProcessorResponse (generating a TypeError)
* The PaymentMethod was not a card (generating a KeyError).

Also switched Cybersource to not fulfill if it cannot find the card
brand.