fix: send stripe waffle flag via url parameter in frontend
django-waffle bins users into a waffle flag experience using a cookie or
URL parameter. If a user does not have that cookie or URL parameter,
django-waffle assigns one randomly using the % distribution set by the
waffle flag.

This caused the Payment MFE to not be consistent in which waffle flag it
uses to communicate with Ecommerce when we enabled the Ecommerce
`enable_stripe_payment_processor` waffle flag for a % of users (as
opposed to a list of users, or for all users).

Unfortunately, it does not look like the cookie is automatically passed
along between Ecommerce and the Payment MFE like we hoped. However, wome
nifty work in 9e9e54630cad055b1599ab7bac4935f5d6b55b6e shows that we
used to pass flags to the frontends via URL parameters.

The Payment MFE already has a
[parser](https://github.com/openedx/frontend-app-payment/blob/4c920f6b20ba98efe80c775df4ae6db3a65a89d5/src/index.jsx#L37-L46)
and
[interceptor](https://github.com/openedx/frontend-app-payment/blob/4c920f6b20ba98efe80c775df4ae6db3a65a89d5/src/index.jsx#L93-L103)
to receive waffle flags this way.

This PR re-enables some of the old functionality to make waffle flags
work.