feat: upgrade edx-drf-extensions and newrelic (#4027)
* fix: test update for edx-drf-extensions 8.11.0

The method generate_unversioned_payload in
edx-drf-extensions now requires a user.id to
function, so that needed to be added to the mock
user.

* feat: upgrade edx-drf-extensions

Upgrade edx-drf-extensions to 8.12.0. This was done
using a temporary constraint and pip-compile without
the --upgrade to keep from upgrading unrelated
requirements.

* feat: upgrade newrelic

Upgrade newrelic to 9.1.0. This was done
using a temporary constraint and pip-compile without
the --upgrade to keep from upgrading unrelated
requirements.