feat: upgrade edx-drf-extensions to 9.0.1
9.0.1 fixes ENABLE_FORGIVING_JWT_COOKIES bug.
This build on the 9.0.0 fix, but applies the fix
to another place where the JWT's LMS user id will
be compared in place of the newly authenticated
service user's id.

For details, see:
https://github.com/openedx/edx-drf-extensions/pull/414

Note that this won't be put into effect until
ENABLE_FORGIVING_JWT_COOKIES is toggled on
separately.

This is part of the rollout of:
https://github.com/edx/edx-arch-experiments/issues/429